requirements.txt
playwright
python-dotenv
google-generativeai